+++ backend/windmill-worker/src/worker.rs
@@ -15,6 +15,8 @@ use sqlx::Pool;
 use sqlx::Postgres;
 use tokio::time::timeout;
 use windmill_common::client::AuthedClient;
+use windmill_common::scripts::hash_to_codebase_id;
+use windmill_common::scripts::is_special_codebase_hash;
 use windmill_common::utils::report_critical_error;
 use windmill_common::utils::retrieve_common_worker_prefix;
 use windmill_common::{
@@ -22,7 +24,6 @@ use windmill_common::{
     apps::AppScriptId,
     cache::{future::FutureCachedExt, ScriptData, ScriptMetadata},
     schema::{should_validate_schema, SchemaValidator},
-    scripts::PREVIEW_IS_TAR_CODEBASE_HASH,
     utils::{create_directory_async, WarnAfterExt},
     worker::{
         make_pull_query, write_file, Connection, HttpClient, MAX_TIMEOUT,
@@ -66,7 +67,7 @@ use windmill_common::{
     error::{self, to_anyhow, Error},
     flows::FlowNodeId,
     jobs::JobKind,
-    scripts::{get_full_hub_script_by_path, ScriptHash, ScriptLang, PREVIEW_IS_CODEBASE_HASH},
+    scripts::{get_full_hub_script_by_path, ScriptHash, ScriptLang},
     utils::StripPath,
     worker::{CLOUD_HOSTED, NO_LOGS, WORKER_CONFIG, WORKER_GROUP},
     DB, IS_READY,
@@ -2387,12 +2388,13 @@ pub async fn handle_queued_job(
             | JobKind::Flow
             | JobKind::FlowDependencies,
             x,
-        ) => match x.map(|x| x.0) {
-            None | Some(PREVIEW_IS_CODEBASE_HASH) | Some(PREVIEW_IS_TAR_CODEBASE_HASH) => Some(
+        ) => if x.map(|x| x.0).is_none_or(|x| is_special_codebase_hash(x)) {
+            Some(
                 cache::job::fetch_preview(conn, &job.id, raw_lock, raw_code, raw_flow.clone())
                     .await?,
-            ),
-            _ => None,
+            ) 
+        } else {
+                None
         },
         _ => None,
     };
@@ -2887,12 +2889,7 @@ async fn handle_code_execution_job(
         ScriptMetadata { language, envs, codebase, schema_validator, schema },
     ) = match job.kind {
         JobKind::Preview => {
-            let codebase = match job.runnable_id.map(|x| x.0) {
-                Some(PREVIEW_IS_CODEBASE_HASH) => Some(job.id.to_string()),
-                Some(PREVIEW_IS_TAR_CODEBASE_HASH) => Some(format!("{}.tar", job.id)),
-                _ => None,
-            };
-
+            let codebase = job.runnable_id.and_then(|x| hash_to_codebase_id(&job.id.to_string(), x.0));
             if codebase.is_none() && job.runnable_id.is_some() {
                 (arc_data, arc_metadata) =
                     cache::script::fetch(conn, job.runnable_id.unwrap()).await?;